+++ jdaviz/configs/imviz/plugins/footprints/tests/test_footprints.py
@@ -1,2 +1,3 @@
 def test_user_api(imviz_helper):
     plugin = imviz_helper.plugins['Footprints']
+    plugin.plugin_opened = True